+++ lib/convert.dart
@@ -501,8 +501,9 @@ Output path: "$outFilePathEx"
 
     _logger.out('Expanding: "$inpPath"$outPath${isExpandContentOnly ? '' : '\n'}');
 
-    // Load the input as a text string
+    // Get permissions and load the input as a text string
 
+    var permissions = inpFile?.statSync();
     var text = (inpFile == null ? stdin.readAsStringSync() : inpFile.readAsStringSync());
 
     // Remove keywords and escape special characters
@@ -649,6 +650,10 @@ Output path: "$outFilePathEx"
 
         Path.fileSystem.file(outFilePath)
           .writeAsStringSync(text);
+
+        if (permissions != null) {
+          Command.chmod(permissions.mode, outFilePath);
+        }
       }
       else {
         tmpFile = Path.fileSystem.file(tmpFilePath);
@@ -661,7 +666,11 @@ Output path: "$outFilePathEx"
           ..deleteIfExistsSync()
           ..writeAsStringSync(text);
 
-        if (Path.equals(inpFilePath, outFilePath)) {
+        if (permissions != null) {
+          Command.chmod(permissions.mode, outFilePath);
+        }
+
+        if (Path.equals(inpFilePath, tmpFile.path)) {
           if (_logger.isVerbose) {
             _logger.verbose('...renaming temp file to $outFilePath');
           }